+++ tests/test_runner_sweep_integration.py
@@ -137,9 +137,9 @@ class TestRunWorkflowRunIdInjection:
         runner = WorkflowRunner(workflow)
         return runner, job
 
-    @patch("srunx.runner._transition_workflow_run")
+    @patch("srunx.runtime.workflow.runner._transition_workflow_run")
     @patch("srunx.db.cli_helpers.create_cli_workflow_run")
-    @patch("srunx.runner.Slurm")
+    @patch("srunx.runtime.workflow.runner.Slurm")
     def test_injected_id_skips_create(
         self,
         mock_slurm_class,
@@ -166,9 +166,9 @@ class TestRunWorkflowRunIdInjection:
         assert mock_transition.called
         assert mock_transition.call_args_list[0].args[0] == injected_id
 
-    @patch("srunx.runner._transition_workflow_run")
+    @patch("srunx.runtime.workflow.runner._transition_workflow_run")
     @patch("srunx.db.cli_helpers.create_cli_workflow_run")
-    @patch("srunx.runner.Slurm")
+    @patch("srunx.runtime.workflow.runner.Slurm")
     def test_no_id_falls_back_to_create(
         self,
         mock_slurm_class,
@@ -189,9 +189,9 @@ class TestRunWorkflowRunIdInjection:
         call_kwargs = mock_slurm.run.call_args.kwargs
         assert call_kwargs["workflow_run_id"] == 999
 
-    @patch("srunx.runner._transition_workflow_run")
+    @patch("srunx.runtime.workflow.runner._transition_workflow_run")
     @patch("srunx.db.cli_helpers.create_cli_workflow_run")
-    @patch("srunx.runner.Slurm")
+    @patch("srunx.runtime.workflow.runner.Slurm")
     def test_create_returning_none_does_not_call_state_service(
         self,
         mock_slurm_class,
